Implemented FoodPlanner Algorithm

+++ MEHR/Other/FoodPlannerQuery.cs
@@ -4,6 +4,6 @@ public record struct FoodPlannerQuery
 (
     float Diversity,
     float ExternalInfluence,
-    double? MaxPriceInEUR,
-    decimal? MaxDistanceInKM
+    decimal? MaxPriceInEUR,
+    double? MaxDistanceInKM
 ) { }
